requirements.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mysql-connector-python==9.0.0
packaging==24.1
pluggy==1.5.0
pyclean==3.0.0
PyMySQL==1.1.1
pytest==8.3.3
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
uuid==1.30
Werkzeug==3.0.4
